Arrumação de views em produtos e problemas com listar pedidos como admin

+++ MIXCOM com LARAVEL 4.0/app/Http/Controllers/MercadoPagoController.php
@@ -34,7 +34,7 @@ class MercadoPagoController extends Controller
             $idpedido = DB::table('pedido_produtos')->max('pedido_id');
         }
         $somapedido = DB::table('pedido_produtos')->where('pedido_id', $idpedido)->sum('valor');
-        $pedido = (float)$somapedido;
+        $pedido = (int)$somapedido;
 
         $compra = DB::table('pedido_produtos')->where('pedido_id', $idpedido)->value('pedido_id');
